V1.1.9
minor bug fixes

+++ app/Http/Livewire/Backend/AcademicYear/AllAcademicYear.php
@@ -6,6 +6,7 @@ use Livewire\Component;
 use App\Models\AcademicYear;
 use Livewire\WithPagination;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
 
 class AllAcademicYear extends Component
 {   
@@ -16,6 +17,19 @@ class AllAcademicYear extends Component
     public $year;
     public $status;
     public $c_id;
+    public $current_id;
+
+    protected function rules()
+    {
+        return [
+            'year' => ['required', 'digits:4', 'integer','min:1900','unique:academic_years,year,'.($this->mode=='edit'? $this->current_id :'')]
+        ];
+    }
+
+    public function updated($propertyName)
+    {
+        $this->validateOnly($propertyName);
+    }
 
     public function resetinput()
     {
@@ -23,6 +37,7 @@ class AllAcademicYear extends Component
        $this->status=null;
        $this->c_id=null;
        $this->search =null;
+       $this->current_id=null;
     }
 
     public function setmode($mode)
@@ -32,14 +47,20 @@ class AllAcademicYear extends Component
  
     public function save()
     {   
-        $validatedData=$this->validate([
-            'year' => 'required|digits:4|integer|min:1900|unique:academic_years,year',
-        ]);
-
+        $validatedData=$this->validate();
         $academicyear= new AcademicYear;
-        $academicyear->year = $validatedData['year'];
-        $academicyear->status = $this->status==1?1:0;
-        $academicyear->save();
+        if($academicyear){
+            $academicyear->year = $validatedData['year'];
+            $academicyear->status = $this->status==1?1:0;
+            $academicyear->save();
+        }
+        else{
+
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->resetinput();
         $this->setmode('all');
         $this->dispatchBrowserEvent('alert',[
@@ -49,29 +70,41 @@ class AllAcademicYear extends Component
     }
 
     public function edit($id)
-    { 
+    {   
+        $this->current_id=$id;
         $academicyear = AcademicYear::find($id);
-        $this->C_id=$academicyear->id;
-        $this->year = $academicyear->year;
-        $this->status = $academicyear->status;
+        if($academicyear)
+        {
+            $this->C_id=$academicyear->id;
+            $this->year = $academicyear->year;
+            $this->status = $academicyear->status;
+        }else{
+
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->setmode('edit');
     }
 
     public function update($id)
     {   
-        $validatedData=$this->validate([
-            'year' => [
-                'required',
-                'digits:4',
-                'integer',
-                'min:1900',
-                Rule::unique('academic_years', 'year')->ignore($this->year, 'year'),
-            ],
-        ]);
+        $validatedData=$this->validate();
         $academicyear = AcademicYear::find($id);
-        $academicyear->year = $validatedData['year'];
-        $academicyear->status = $this->status==1?'1':'0';
-        $academicyear->update();
+        if($academicyear)
+        {
+            $academicyear->year = $validatedData['year'];
+            $academicyear->status = $this->status==1?'1':'0';
+            $academicyear->update();
+
+        }else{
+
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->resetinput();
         $this->setmode('all');
         $this->dispatchBrowserEvent('alert',[
@@ -83,7 +116,17 @@ class AllAcademicYear extends Component
     public function delete($id)
     { 
         $academicyear = AcademicYear::find($id);
-        $academicyear->delete();
+        if($academicyear)
+        {
+            $academicyear->delete();
+            
+        }else{
+
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->setmode('all');
         $this->dispatchBrowserEvent('alert',[
             'type'=>'success',
@@ -94,6 +137,6 @@ class AllAcademicYear extends Component
     public function render()
     {   
         $academicyear=AcademicYear::where('year', 'like', '%'.$this->search.'%')->orderBy('year', 'DESC')->paginate($this->per_page);
-        return view('livewire.backend.academic-year.all-academic-year',compact('academicyear'))->extends('layouts.admin')->section('admin');
+        return view('livewire.backend.academic-year.all-academic-year',compact('academicyear'))->extends('layouts.admin.admin')->section('admin');
     }
 }
