+++ app/Http/Livewire/Backend/Admin/AllAdmin.php
@@ -30,18 +30,21 @@ class AllAdmin extends Component
     public $password_confirmation;
     public $photo;
     public $photoold;
-
-
+    public $current_id;
+    
     protected function rules()
-    {
+    {   
+        $passwordRules = $this->mode == 'add' ? ['required', 'same:password_confirmation', 'string', 'min:8', 'max:255'] : [];
+
         return [
+
+            'password' => $passwordRules,
             'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255','unique:admins,email,'.($this->mode=='edit'? Auth::user()->id :'')], 
-            'password' => [($this->password=null? 'nullable' : 'required'),'required','string', 'min:8', 'max:255'],
-            'mobile' => ['nullable', 'numeric', 'digits:10','unique:admins,mobile,'.($this->mode=='edit'? Auth::user()->id :'')], 
+            'email' => ['required', 'string', 'email', 'max:255','unique:admins,email,'.($this->mode=='edit'? $this->current_id :'')], 
+            'mobile' => ['nullable', 'numeric', 'digits:10','unique:admins,mobile,'.($this->mode=='edit'? $this->current_id :'')], 
             'photo' => ['nullable', 'image', 'mimes:jpg,png,jpeg', 'max:1024'],
             'role_id' => ['required', 'integer',],
-        ];
+        ];  
     }
 
     public function updated($propertyName)
@@ -49,20 +52,20 @@ class AllAdmin extends Component
         $this->validateOnly($propertyName);
     }
 
- 
     public function resetinput()
     {
         $this->c_id=null;
         $this->name=null;
         $this->email=null;
-        $this->password=null;
         $this->mobile=null;
         $this->status=null;
         $this->role_id=null;
+        $this->password=null;
         $this->password_confirmation=null;
         $this->photo=null;
         $this->photoold=null;
         $this->search =null;
+        $this->current_id=null;
     }
 
     public function setmode($mode)
@@ -74,20 +77,30 @@ class AllAdmin extends Component
     {  
         $validatedData =    $this->validate();  
         $admin= new Admin;
-        $admin->role_id = $validatedData['role_id'];
-        $admin->name = $validatedData['name'];
-        $admin->email= $validatedData['email'];
-        $admin->password= Hash::make($validatedData['password']);
-        $admin->mobile= $validatedData['mobile'];
-        $admin->status = $this->status==1?'1':'0';
-        if($this->photo)
-        {   $path='uploads/profile/admin/photo/';
-            $FileName = 'user-'.now()->format('Y-m-d').'.'.$this->photo->getClientOriginalExtension();
-            $this->photo->storeAs($path,$FileName,'public');
-            $admin->photo='storage/'.$path.$FileName ;
-            $this->reset('photo');
+        if($admin)
+        {
+            $admin->role_id = $validatedData['role_id'];
+            $admin->name = $validatedData['name'];
+            $admin->email= $validatedData['email'];
+            $admin->password= Hash::make($validatedData['password']);
+            $admin->mobile= $validatedData['mobile'];
+            $admin->status = $this->status==1?'1':'0';
+            if($this->photo)
+            {   $path='uploads/profile/admin/photo/';
+                $FileName = 'user-'.now()->format('Y-m-d').'.'.$this->photo->getClientOriginalExtension();
+                $this->photo->storeAs($path,$FileName,'public');
+                $admin->photo='storage/'.$path.$FileName ;
+                $this->reset('photo');
+            }
+            $admin->save();
+            
+        }else{
+
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
         }
-        $admin->save();
         $this->resetinput();
         $this->setmode('all');
         $this->dispatchBrowserEvent('alert',[
@@ -98,14 +111,24 @@ class AllAdmin extends Component
 
     public function edit($id)
     {   
+        $this->current_id=$id;
         $admin = Admin::find($id);
-        $this->C_id=$admin->id;
-        $this->role_id=$admin->role_id;
-        $this->name=$admin->name;
-        $this->email=$admin->email;
-        $this->mobile=$admin->mobile;
-        $this->photoold=$admin->photo;
-        $this->status =$admin->status;
+        if($admin)
+        {
+            $this->C_id=$admin->id;
+            $this->role_id=$admin->role_id;
+            $this->name=$admin->name;
+            $this->email=$admin->email;
+            $this->mobile=$admin->mobile;
+            $this->photoold=$admin->photo;
+            $this->status =$admin->status;
+        }else{
+
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->setmode('edit');
     }
 
@@ -113,19 +136,29 @@ class AllAdmin extends Component
     {   
         $validatedData =  $this->validate();  
         $admin = Admin::find($id);
-        $admin->role_id = $validatedData['role_id'];
-        $admin->name = $validatedData['name'];
-        $admin->email= $validatedData['email'];
-        $admin->mobile= $validatedData['mobile'];
-        $admin->status = $this->status==1?'1':'0';
-        if($this->photo)
-        {   $path='uploads/profile/admin/photo/';
-            $FileName = 'user-'.now()->format('Y-m-d').'.'.$this->photo->getClientOriginalExtension();
-            $this->photo->storeAs($path,$FileName,'public');
-            $admin->photo='storage/'.$path.$FileName ;
-            $this->reset('photo');
+        if($admin)
+        {
+            $admin->role_id = $validatedData['role_id'];
+            $admin->name = $validatedData['name'];
+            $admin->email= $validatedData['email'];
+            $admin->mobile= $validatedData['mobile'];
+            $admin->status = $this->status==1?'1':'0';
+            if($this->photo)
+            {   $path='uploads/profile/admin/photo/';
+                $FileName = 'user-'.now()->format('Y-m-d').'.'.$this->photo->getClientOriginalExtension();
+                $this->photo->storeAs($path,$FileName,'public');
+                $admin->photo='storage/'.$path.$FileName ;
+                $this->reset('photo');
+            }
+            $admin->update();
+
+        }else{
+
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
         }
-        $admin->update();
         $this->resetinput();
         $this->dispatchBrowserEvent('alert',[
             'type'=>'success',
@@ -137,11 +170,21 @@ class AllAdmin extends Component
     public function delete($id)
     { 
         $admin = Admin::find($id);
-        if($admin->photo)
+        if($admin)
         {
-            File::delete($admin->photo);
+            if($admin->photo)
+            {
+                File::delete($admin->photo);
+            }
+            $admin->delete();
+
+        }else{
+
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
         }
-        $admin->delete();
         $this->setmode('all');
         $this->dispatchBrowserEvent('alert',[
             'type'=>'success',
@@ -153,7 +196,7 @@ class AllAdmin extends Component
     {   
         $roles=Role::where('status',0)->orderBy('role', 'ASC')->get();
         $admins=Admin::where('name', 'like', '%'.$this->search.'%')->orderBy('name', 'ASC')->paginate($this->per_page);
-        return view('livewire.backend.admin.all-admin',compact('admins','roles'))->extends('layouts.admin')->section('admin');
+        return view('livewire.backend.admin.all-admin',compact('admins','roles'))->extends('layouts.admin.admin')->section('admin');
     }
 
 }
