+++ app/Http/Livewire/Backend/Building/AllBuilding.php
@@ -20,18 +20,22 @@ class AllBuilding extends Component
     public $status;
     public $hostel_id;
     public $c_id;
+    public $current_id;
 
-
-    protected $rules = [
-        'name' => ['required','string'],
-        'hostel_id' => ['required','integer'],
-    ];
+    protected function rules()
+    {
+        return [
+            'name' => ['required', 'string', 'max:255','unique:buildings,name,'.($this->mode=='edit'? $this->current_id :'')],
+            'hostel_id' => ['required','integer'],
+        ];
+    }
 
     public function updated($propertyName)
     {
         $this->validateOnly($propertyName);
     }
- 
+
+
     public function resetinput()
     {
         $this->hostel_id=null;
@@ -40,6 +44,7 @@ class AllBuilding extends Component
         $this->status=null;
         $this->c_id=null;
         $this->building_name =null;
+        $this->current_id=null;
     }
 
     public function setmode($mode)
@@ -51,10 +56,17 @@ class AllBuilding extends Component
     {
         $validatedData =  $this->validate();
         $building= new Building;
-        $building->name = $validatedData['name'];
-        $building->hostel_id = $validatedData['hostel_id'];
-        $building->status = $this->status==1?1:0;
-        $building->save();
+        if($building){
+            $building->name = $validatedData['name'];
+            $building->hostel_id = $validatedData['hostel_id'];
+            $building->status = $this->status==1?1:0;
+            $building->save();
+        }else{
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->resetinput();
         $this->setmode('all');
         $this->dispatchBrowserEvent('alert',[
@@ -64,12 +76,20 @@ class AllBuilding extends Component
     }
 
     public function edit($id)
-    { 
+    {   
+        $this->current_id=$id;
         $building = Building::find($id);
-        $this->C_id=$building->id;
-        $this->status = $building->status;
-        $this->name = $building->name;
-        $this->hostel_id = $building->hostel_id;
+        if($building){
+            $this->C_id=$building->id;
+            $this->status = $building->status;
+            $this->name = $building->name;
+            $this->hostel_id = $building->hostel_id;
+        }else{
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->setmode('edit');
     }
 
@@ -77,10 +97,17 @@ class AllBuilding extends Component
     {   
         $validatedData =  $this->validate();
         $building = Building::find($id);
-        $building->name = $validatedData['name'];
-        $building->hostel_id = $validatedData['hostel_id'];
-        $building->status = $this->status==1?1:0;
-        $building->update();
+        if($building){
+            $building->name = $validatedData['name'];
+            $building->hostel_id = $validatedData['hostel_id'];
+            $building->status = $this->status==1?1:0;
+            $building->update();
+        }else{
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->resetinput();
         $this->setmode('all');
         $this->dispatchBrowserEvent('alert',[
@@ -92,7 +119,14 @@ class AllBuilding extends Component
     public function delete($id)
     { 
         $building = Building::find($id);
-        $building->delete();
+        if($building){
+            $building->delete();
+        }else{
+            $this->dispatchBrowserEvent('alert',[
+                'type'=>'error',
+                'message'=>"Something Went Wrong!!"
+            ]);
+        }
         $this->setmode('all');
         $this->dispatchBrowserEvent('alert',[
             'type'=>'success',
@@ -104,13 +138,13 @@ class AllBuilding extends Component
     {   
         $hostels=Hostel::where('status',0)->orderBy('name',"ASC")->get();
 
-        $query =Building::orderBy('name', 'ASC');      
+        $query = Building::orderBy('name', 'ASC');
         if ($this->hostel_name) {
-            $hostelIds = Hostel::where('status', 0)->where('name', 'like', '%' . $this->hostel_name. '%')->pluck('id');
-            $query->whereIn('hostel_id', $hostelIds);
+            $query->whereHas('hostel', function ($query) {
+                $query->where('status', 0)->where('name', 'like', '%' . $this->hostel_name . '%');
+            });
         }
-        $building = $query->where('name', 'like', '%'.$this->building_name.'%')->paginate($this->per_page);
-        
-        return view('livewire.backend.building.all-building',compact('building','hostels'))->extends('layouts.admin')->section('admin');
+        $building = $query->where('name', 'like', '%' . $this->building_name . '%')->with('hostel')->paginate($this->per_page);
+        return view('livewire.backend.building.all-building',compact('building','hostels'))->extends('layouts.admin.admin')->section('admin');
     }
 }
